Don't automatically assign port numbers. Refactor.
Automatically assigning port numbers causes a lot of problems when
running apache commands. And there is no (easy) way to map port
numbers back to the vhost they were originally assigned to. So that
feature is gone.

The vhost information is now wrapped up in a class, which does the vhost
file parsing in the constructor. Makes the code a little cleaner.

Added some docstrings and more detailed comments.

No longer uses the "base.conf" config, but instead, the main apache
config. That means certain directives in the main apache config need to
be wrapped in <IfDefine>, like:

    <IfDefine !httpdmulti>
        Listen 80
    </IfDefine>